+++ src/main/java/com/ggapp/dao/repository/mongo/NewsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NewsRepository extends MongoRepository<News, Integer> {
+public interface NewsRepository extends MongoRepository<News, Long> {
 }
