+++ src/main/java/com/ggapp/dao/repository/mongo/SessionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SessionRepository extends MongoRepository<Session, Integer> {
+public interface SessionRepository extends MongoRepository<Session, Long> {
 }
