+++ src/main/java/com/ggapp/dao/repository/mongo/UserRepository.java
@@ -13,7 +13,7 @@ import java.util.Optional;
  * @author Tran Minh Truyen
  */
 @Repository
-public interface UserRepository extends MongoRepository<User, Integer> {
+public interface UserRepository extends MongoRepository<User, Long> {
     Optional<List<User>> findUserByFirstNameContainingOrLastNameContaining(String firstName, String lastName);
     Optional<User> findByEmailEqualsIgnoreCase(String email);
 }
