+++ src/main/java/com/ggapp/services/CartService.java
@@ -7,12 +7,12 @@ import com.ggapp.common.exception.ApplicationException;
  * @author Tran Minh Truyen
  */
 public interface CartService {
-	CartResponse createCart(int customerId, int productId, long productAmount) throws ApplicationException;
-	CartResponse getCartById(int id) throws ApplicationException;
-	CartResponse updateProductAmountInCart(int customerId, int productId, long amount) throws ApplicationException;
-	CartResponse deleteCart(int id);
-	boolean deleteCartAfterCreateOrder(int id);
-	CartResponse addProductToCart(int customerId, int productId, long productAmount) throws ApplicationException;
-	CartResponse removeProductFromCart(int customerId, int productId);
-	boolean isCartExists(int customerId);
+	CartResponse createCart(Long customerId, Long productId, long productAmount) throws ApplicationException;
+	CartResponse getCartById(Long id) throws ApplicationException;
+	CartResponse updateProductAmountInCart(Long customerId, Long productId, long amount) throws ApplicationException;
+	CartResponse deleteCart(Long id);
+	boolean deleteCartAfterCreateOrder(Long id);
+	CartResponse addProductToCart(Long customerId, Long productId, long productAmount) throws ApplicationException;
+	CartResponse removeProductFromCart(Long customerId, Long productId);
+	boolean isCartExists(Long customerId);
 }
